[SMIT]final changes

+++ src/resource/review/model.py
@@ -1,14 +1,15 @@
 from database.database import Base
-from sqlalchemy import Column,String,DateTime,ForeignKey,Integer
+from sqlalchemy import Column, String, DateTime, ForeignKey, Float
 from datetime import datetime
 from src.resource.product.model import Product
 from src.resource.user.model import User
 
+
 class Review(Base):
-    __tablename__ = 'reviews'
+    __tablename__ = "reviews"
     id = Column(String, primary_key=True)
     content = Column(String, nullable=False)
-    rating = Column(Integer, nullable=False)
+    rating = Column(Float, nullable=False)
     product_id = Column(String, ForeignKey(Product.id), nullable=False)
     user_id = Column(String, ForeignKey(User.id), nullable=False)
     created_at = Column(DateTime, default=datetime.now())
