+++ src/resource/review/schema.py
@@ -1,11 +1,13 @@
 from pydantic import BaseModel
 from typing import Optional
 
+
 class ReviewRequest(BaseModel):
-    product_id : str
-    content : str
-    rating : int
+    product_id: str
+    content: str
+    rating: float
+
 
 class ReviewUpdate(BaseModel):
-    content : Optional[str]= None
-    rating :Optional[int]= None
+    content: Optional[str] = None
+    rating: Optional[float] = None
